restructures files, created utils, added TSP code
