+++ data/Portland Bus Route Locations- Grocery Stores.csv
@@ -1,0 +1,11 @@
+WKT,name,description
+"POINT (-70.2646208 43.66181479999999)",Trader Joe's,Grocery
+"POINT (-70.2584734 43.6634486)",Whole Foods Market,Grocery
+"POINT (-70.2681516 43.6639522)",Hannaford,Grocery
+"POINT (-70.252803 43.6609941)",Portland Food Co-op,Grocery
+"POINT (-70.2977049 43.6575062)",Shaw's,Grocery
+"POINT (-70.2802394 43.65224480000001)",Save A Lot,Grocery
+"POINT (-70.2581426 43.6598119)",Peace Food Market,Grocery
+"POINT (-70.3196408 43.7018373)",Hannaford,Grocery
+"POINT (-70.2880275 43.7024145)",Shaw's,Grocery
+"POINT (-70.2496714 43.6602986)",Micucci Grocery Store,Grocery
