+++ data/Portland Bus Route Locations- Healthcare Facilities.csv
@@ -1,0 +1,12 @@
+WKT,name,description
+"POINT (-70.27838009999999 43.6535614)",Greater Portland Health - 934 Congress,Healthcare
+"POINT (-70.2466965 43.6877174)",Martin's Point Health Care Center,Healthcare
+"POINT (-70.27428119999999 43.6560044)",Portland Community Health Center,Healthcare
+"POINT (-70.262609 43.679923)",Seaside Healthcare,Healthcare
+"POINT (-70.26370899999999 43.6556669)",Portland Community Free Clinic,Healthcare
+"POINT (-70.27615229999999 43.6531296)",Maine Medical Center,Healthcare
+"POINT (-70.2820771 43.6482809)",Northern Light Mercy Hospital,Healthcare
+"POINT (-70.261062 43.6641234)",ConvenientMD Urgent Care,Healthcare
+"POINT (-70.25554040000002 43.6351995)",AFC Urgent Care South Portland,Healthcare
+"POINT (-70.3171267 43.6350152)",Concentra Urgent Care,Healthcare
+"POINT (-70.3250206 43.6383648)",InterMed Urgent Care,Healthcare
