--- locations data/Portland Bus Route Locations- Schools.csv
@@ -1,17 +1,0 @@
-WKT,name,description
-"POINT (-70.25365099999999 43.671422)",East End Community School,School
-"POINT (-70.2964921 43.6714847)",Deering High School,School
-"POINT (-70.2590003 43.6592521)",Portland High School,School
-"POINT (-70.29555719999999 43.6729832)",Longfellow Elementary School,School
-"POINT (-70.29444 43.697413)",Casco Bay High School,School
-"POINT (-70.25963 43.6843118)",Presumpscot Elementary School,School
-"POINT (-70.2851373 43.7081343)",Harrison Lyseth Elementary School,School
-"POINT (-70.26666039999999 43.6503481)",Reiche Community School,School
-"POINT (-70.28378959999999 43.6771399)",Ocean Avenue Elementary School,School
-"POINT (-70.28803979999999 43.7067378)",Lyman Moore Middle School,School
-"POINT (-70.2757206 43.6619019)",University of Southern Maine,School
-"POINT (-70.2290826 43.6469528)",Southern Maine Community College,School
-"POINT (-70.2466249 43.6615206)",The Roux Institute at Northeastern University,School
-"POINT (-70.29641529999999 43.683423)",University of New England - Westbrook College Campus,School
-"POINT (-70.2822578 43.6723699)",University of New England,School
-"POINT (-70.2605967 43.6556992)",Maine College of Art & Design,School
